Change parse table format (version 7):
Character classes now explicitly store whether they contain EOF,
instead of using an integer (used to be 256) for this.
Examples: character-class([eof])  character-class([range(65,90),97,eof])

The ParseTableReader is equipped to read parse tables
of both versions 6 and 7 for backwards compatibility.
